Address API route review feedback
