+++ apps/api/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@fantasy-sumo/db": "workspace:*",
     "@fantasy-sumo/domain": "workspace:*",
-    "fastify": "^5.6.2"
+    "fastify": "^5.6.2",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
     "@types/node": "^24.10.1",
